# api.yaml
openapi: 3.0.3
info:
  title: API ALiVE
  description: |-
    Cette interface est la documentation de l'API ALiVE

  contact:
    email: gaspard.vandecasteele@student.junia.com
  version: "1.0"
externalDocs:
  description: Retrouver le jeu ALiVE
  url: http://exemple.com
servers:
  - url: http://localhost:8080
tags:
  - name: map
    description: accès au map
  - name: object
    description: accès aux objets
  - name: place
    description: accès aux places
  - name: room
    description: accès aux pièces

paths:
  /api/room:
    get:
      tags:
        - room
      summary: Récupère toutes les pièces
      description: Récupère toutes les pièces
      operationId: getRooms
      responses:
        '200':
          $ref: '#/components/responses/RoomSucces'

        '404':
          description: Aucune pièces trouvées
  /api/room/{room_uuid}:
    get:
      tags:
        - room
      summary: Récupère une pièce précise
      description: Récupère une pièce précise grâce a sont uuid
      parameters:
        - $ref: '#/components/parameters/RoomParam'
      responses:
        '200':
          $ref: '#/components/responses/RoomSucces'          

        '400':
          description: Invalid input
    patch:
      tags:
        - room
      summary: Modifie une pièce
      description: modifie une pièce deja existante
      parameters:
        - $ref: '#/components/parameters/RoomParam'
      requestBody:
        $ref: '#/components/requestBodies/RoomBody'
      responses:
        '200':
          $ref: '#/components/responses/RoomSucces'          

        '400':
          description: Invalid input
    delete:
      tags:
        - room
      summary: Supprime une pièce
      description: Supprime une pièce précise
      parameters:
        - $ref: '#/components/parameters/RoomParam'
      responses:
        '200':
          $ref: '#/components/responses/DeleteSucces'
  /api/place/{place_uuid}/room:
    post:
      tags:
        - place
        - room
      summary: Ajoute une nouvelle pièce
      description: Ajoute une nouvelle pièce dans une place
      operationId: addRoom
      parameters:
        - $ref: '#/components/parameters/PlaceParam'
      requestBody:
        $ref: '#/components/requestBodies/RoomBody' 
            
      responses:
        '200':
          $ref: '#/components/responses/RoomSucces'          

        '400':
          description: Invalid input
  /api/map:
    get:
      tags:
        - map
      summary: Récupère toutes les cartes
      description: Récupère toutes les cartes et leurs informations
      responses:
        '200':
          description: Une map à été créé, modifié ou récupéré
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    $ref: '#/components/schemas/Status'
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/WholeMap'
    post:
      tags:
        - map
      summary: Créé une nouvelle carte
      description: Créé une nouvelle carte et retourne ses informations
      requestBody:
        $ref: '#/components/requestBodies/MapBody'
      responses:
        '200':
          $ref: '#/components/responses/MapSucces'

  /api/object:
    post:
      tags:
        - object
      summary: Créer un objet
      description: Création d'un objet et du fichier d'image associé
      requestBody:
        $ref: '#/components/requestBodies/CreateObject'

components:
  schemas:
    CreateObject:
      type: object
      properties:
        name:
          type: string
          example: 'test'
        description:
          type: string
          example: 'test'
        isTool:
          type: boolean
          example: true
        fileName:
          type: string
          example: 'test.png'
        fileData:
          type: string
          example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABgAAAAYCAYAAADgdz34AAAABHNCSVQICAgIfAhkiAAAAAlwSFlzAAAApgAAAKYB3X3/OAAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAAANCSURBVEiJtZZPbBtFFMZ/M7ubXdtdb1xSFyeilBapySVU8h8OoFaooFSqiihIVIpQBKci6KEg9Q6H9kovIHoCIVQJJCKE1ENFjnAgcaSGC6rEnxBwA04Tx43t2FnvDAfjkNibxgHxnWb2e/u992bee7tCa00YFsffekFY+nUzFtjW0LrvjRXrCDIAaPLlW0nHL0SsZtVoaF98mLrx3pdhOqLtYPHChahZcYYO7KvPFxvRl5XPp1sN3adWiD1ZAqD6XYK1b/dvE5IWryTt2udLFedwc1+9kLp+vbbpoDh+6TklxBeAi9TL0taeWpdmZzQDry0AcO+jQ12RyohqqoYoo8RDwJrU+qXkjWtfi8Xxt58BdQuwQs9qC/afLwCw8tnQbqYAPsgxE1S6F3EAIXux2oQFKm0ihMsOF71dHYx+f3NND68ghCu1YIoePPQN1pGRABkJ6Bus96CutRZMydTl+TvuiRW1m3n0eDl0vRPcEysqdXn+jsQPsrHMquGeXEaY4Yk4wxWcY5V/9scqOMOVUFthatyTy8QyqwZ+kDURKoMWxNKr2EeqVKcTNOajqKoBgOE28U4tdQl5p5bwCw7BWquaZSzAPlwjlithJtp3pTImSqQRrb2Z8PHGigD4RZuNX6JYj6wj7O4TFLbCO/Mn/m8R+h6rYSUb3ekokRY6f/YukArN979jcW+V/S8g0eT/N3VN3kTqWbQ428m9/8k0P/1aIhF36PccEl6EhOcAUCrXKZXXWS3XKd2vc/TRBG9O5ELC17MmWubD2nKhUKZa26Ba2+D3P+4/MNCFwg59oWVeYhkzgN/JDR8deKBoD7Y+ljEjGZ0sosXVTvbc6RHirr2reNy1OXd6pJsQ+gqjk8VWFYmHrwBzW/n+uMPFiRwHB2I7ih8ciHFxIkd/3Omk5tCDV1t+2nNu5sxxpDFNx+huNhVT3/zMDz8usXC3ddaHBj1GHj/As08fwTS7Kt1HBTmyN29vdwAw+/wbwLVOJ3uAD1wi/dUH7Qei66PfyuRj4Ik9is+hglfbkbfR3cnZm7chlUWLdwmprtCohX4HUtlOcQjLYCu+fzGJH2QRKvP3UNz8bWk1qMxjGTOMThZ3kvgLI5AzFfo379UAAAAASUVORK5CYII="

    Status:
      type: string
      description: le status de la réponse API
      example: success
      enum:
        - success
        - error
    
    uuid:
      type: string
      example: 78a7eb29967311edaa340242ac1b0003
    
    Uuid:
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/uuid'
    
    Coordinates:
      type: object
      properties:
        x:
          type: integer
          example: 176
        y:
          type: integer
          example: 34
    Room:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: salle de classe
    Place:
      type: object
      properties:
        name:
          type: string
          example: école d'agriculture
        coordinates:
          $ref: '#/components/schemas/Coordinates'
    Map:
      type: object
      properties:
        name:
          type: string
          example: Région nord est
    Day:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    Dialogue:
      type: object
      properties:
        description:
          type: string
    Sentence:
      type: object
      properties:
        content: 
          type: string
        color:
          type: string
    Characters:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
    Object:
      type: object
      properties:
        name:
          type: string
          example: Arrosoir
        description:
          type: string
          example: permet d'arroser
        isTool:
          type: integer
          example: 1
    Door:
      type: object
      properties:
        coordinates:
          $ref: '#/components/schemas/Coordinates'
        distinationDoor:
          $ref: '#/components/schemas/uuid'

    RoomUuid:
      allOf:
        - $ref: '#/components/schemas/Room'
        - $ref: '#/components/schemas/Uuid'
    PlaceUuid:
      allOf:
        - $ref: '#/components/schemas/Place'
        - $ref: '#/components/schemas/Uuid'
    MapUuid:
      allOf:
        - $ref: '#/components/schemas/Map'
        - $ref: '#/components/schemas/Uuid'
    DayUuid:
      allOf:
        - $ref: '#/components/schemas/Day'
        - $ref: '#/components/schemas/Uuid'
    DialogueUuid:
      allOf:
        - $ref: '#/components/schemas/Dialogue'
        - $ref: '#/components/schemas/Uuid'
    SentenceUuid:
      allOf:
        - $ref: '#/components/schemas/Sentence'
        - $ref: '#/components/schemas/Uuid'
    CharactersUuid:
      allOf:
        - $ref: '#/components/schemas/Characters'
        - $ref: '#/components/schemas/Uuid'
    ObjectUuid:
      allOf:
        - $ref: '#/components/schemas/Object'
        - $ref: '#/components/schemas/Uuid'
    DoorUuid:
      allOf:
        - $ref: '#/components/schemas/Door'
        - $ref: '#/components/schemas/Uuid'
    WholeMap:
      allOf:
        - $ref: '#/components/schemas/MapUuid'
        - type: object
          properties:
            places:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/PlaceUuid'
                  - type: object
                    properties:
                      rooms:
                        type: array
                        items:
                          allOf:
                           - $ref: '#/components/schemas/RoomUuid'
                           -  type: object
                              properties:
                                objects:
                                  type: array
                                  items:
                                    allOf:
                                      - $ref: '#/components/schemas/ObjectUuid'
                                doors:
                                  type: array
                                  items:
                                   allOf:
                                    - $ref: '#/components/schemas/DoorUuid'
          
        
    
  
  responses:
    DeleteSucces:
      description: Un objet quelconque à été supprimé
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                $ref: '#/components/schemas/Status'
              content:
                $ref: '#/components/schemas/Uuid' 
    RoomSucces:
      description: Une room à été créé, modifié ou récupéré
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                $ref: '#/components/schemas/Status'
              content:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Room'
                    - $ref: '#/components/schemas/Uuid'
    
    MapSucces:
      description: Une map à été créé, modifié ou récupéré
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                $ref: '#/components/schemas/Status'
              content:
                type: array
                items:
                  $ref: '#/components/schemas/MapUuid'
  requestBodies:
    CreateObject:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateObject'
    RoomBody:
      required: true
      description: |- 
          Ajoute ou modifie une pièce
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Room'
    MapBody:
      required: true
      description: |- 
          Ajoute ou modifie une map
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Map'
  parameters:
    RoomParam:
      in: path
      name: room_uuid
      schema:
        $ref: '#/components/schemas/uuid'
      required: true
      description: uuid d'une room
    PlaceParam:
      in: path
      name: place_uuid
      schema:
        $ref: '#/components/schemas/uuid'
      description: l'uuid de la place
      required: true
  
  # securitySchemes: